--- my_app/trading/api_start.py
@@ -1,1 +1,0 @@
-print("api 실행")
